feat: refactor with neutral Next layout
