+++ src/app/(default)/accessibilite/page.tsx
@@ -4,7 +4,7 @@ import { type Metadata } from "next";
 import { H1 } from "@/dsfr/base/typography";
 import { anchorHeadingMDXComponents } from "@/mdx-components";
 
-import { sharedMetadata } from "../shared-metadata";
+import { sharedMetadata } from "../../shared-metadata";
 
 const title = "Déclaration d'accessibilité";
 const url = "/accessibilite";
@@ -22,10 +22,10 @@ export const metadata: Metadata = {
 };
 
 const Accessibilite = () => (
-  <div className="col-start-2">
+  <>
     <H1>{title}</H1>
     <AccessibiliteContent components={anchorHeadingMDXComponents} />
-  </div>
+  </>
 );
 
 export default Accessibilite;
