+++ src/app/(default)/layout.tsx
@@ -1,7 +1,7 @@
 import { type PropsWithChildren } from "react";
 
 const SimulationLayout = ({ children }: PropsWithChildren) => {
-  return <div className="col-start-2 max-w-[1000px] mt-8 justify-self-center">{children}</div>;
+  return <div className="col-start-2 grid grid-cols-1 mt-6 md:mt-8">{children}</div>;
 };
 
 export default SimulationLayout;
