+++ src/app/(default)/politique-des-cookies/page.tsx
