+++ src/app/(default)/stats/content.tsx
